Did initial port to Gamebuino Meta
Game now compiles and runs on the Meta, although I disabled saving
or loading highscores and sound.  I also had to add a call to
display.clear() because the screen did not clear between updates.

There are a few other small problems (such as the protected cities
not displaying correctly between waves), but mostly all that needs
to be done is to fix sound and highscores, and then add colors and
some meta-specific features.