Added backlights when explosions appear
I decided to take advantage of the LEDs on the back of the Meta,
so I have them display the current explosion color when one is
onscreen.  I had to spend some time tweaking the lights because
they are very bright, and I wanted to avoid too much flashing.
Even now, it seems a little bit intense.